+++ front/src/dashboard/slips/act/RouteSlipsAct.tsx
@@ -4,13 +4,21 @@ import { useQuery, gql } from "@apollo/client";
 import { Query } from "generated/graphql/types";
 import { formSearchResultFragment } from "common/fragments";
 import { Breadcrumb, BreadcrumbItem } from "common/components";
-import { FormSearchResultTable } from "../FormSearchResultTable";
+import { FormSearchResultTable } from "../../FormSearchResultTable";
 
-export const SEARCH_ARCHIVES = gql`
+export const SEARCH_ACTS = gql`
   query SearchFollows($siret: String!) {
     searchForms(
       siret: $siret
-      status: ["PROCESSED", "NO_TRACEABILITY", "REFUSED"]
+      status: [
+        "SENT"
+        "RECEIVED"
+        "ACCEPTED"
+        "TEMP_STORED"
+        "TEMP_STORER_ACCEPTED"
+        "RESENT"
+      ]
+      waitingForMe: true
     ) {
       ...FormSearchResultFragment
     }
@@ -18,9 +26,9 @@ export const SEARCH_ARCHIVES = gql`
   ${formSearchResultFragment}
 `;
 
-export function RouteSlipsHistory() {
+export function RouteSlipsAct() {
   const { siret } = useParams<{ siret: string }>();
-  const { data } = useQuery<Pick<Query, "searchForms">>(SEARCH_ARCHIVES, {
+  const { data } = useQuery<Pick<Query, "searchForms">>(SEARCH_ACTS, {
     variables: {
       siret,
     },
@@ -33,7 +41,7 @@ export function RouteSlipsHistory() {
     <>
       <Breadcrumb>
         <BreadcrumbItem>Mes bordereaux</BreadcrumbItem>
-        <BreadcrumbItem>Archives</BreadcrumbItem>
+        <BreadcrumbItem>Pour action</BreadcrumbItem>
       </Breadcrumb>
       <FormSearchResultTable searchResults={data?.searchForms ?? []} />
     </>
