+++ front/src/dashboard/slips/act/index.ts
@@ -1,0 +1,1 @@
+export * from "./RouteSlipsAct";
