+++ front/src/dashboard/slips/drafts/RouteSlipsDrafts.tsx
@@ -4,35 +4,20 @@ import { useQuery, gql } from "@apollo/client";
 import { Query } from "generated/graphql/types";
 import { formSearchResultFragment } from "common/fragments";
 import { Breadcrumb, BreadcrumbItem } from "common/components";
-import { FormSearchResultTable } from "../FormSearchResultTable";
+import { FormSearchResultTable } from "../../FormSearchResultTable";
 
-export const SEARCH_FOLLOWS = gql`
-  query SearchFollows($siret: String!) {
-    searchForms(
-      siret: $siret
-      status: [
-        "SEALED"
-        "SENT"
-        "RECEIVED"
-        "ACCEPTED"
-        "TEMP_STORED"
-        "TEMP_STORER_ACCEPTED"
-        "RESEALED"
-        "RESENT"
-        "AWAITING_GROUP"
-        "GROUPED"
-      ]
-      waitingForMe: false
-    ) {
+export const SEARCH_DRAFTS = gql`
+  query SearchDrafts($siret: String!) {
+    searchForms(siret: $siret, status: ["DRAFT"]) {
       ...FormSearchResultFragment
     }
   }
   ${formSearchResultFragment}
 `;
 
-export function RouteSlipsFollow() {
+export function RouteSlipsDrafts() {
   const { siret } = useParams<{ siret: string }>();
-  const { data } = useQuery<Pick<Query, "searchForms">>(SEARCH_FOLLOWS, {
+  const { data } = useQuery<Pick<Query, "searchForms">>(SEARCH_DRAFTS, {
     variables: {
       siret,
     },
@@ -45,7 +30,7 @@ export function RouteSlipsFollow() {
     <>
       <Breadcrumb>
         <BreadcrumbItem>Mes bordereaux</BreadcrumbItem>
-        <BreadcrumbItem>Suivi</BreadcrumbItem>
+        <BreadcrumbItem>Brouillons</BreadcrumbItem>
       </Breadcrumb>
       <FormSearchResultTable searchResults={data?.searchForms ?? []} />
     </>
