+++ front/src/dashboard/slips/drafts/index.ts
@@ -1,0 +1,1 @@
+export * from "./RouteSlipsDrafts";
