+++ front/src/dashboard/slips/follow/index.ts
@@ -1,0 +1,1 @@
+export * from "./RouteSlipsFollow";
