+++ front/src/dashboard/slips/history/index.ts
@@ -1,0 +1,1 @@
+export * from "./RouteSlipsHistory";
