+++ front/src/dashboard/slips/index.ts
@@ -1,5 +1,5 @@
-export * from "./RouteSlipsAct";
-export * from "./RouteSlipsDrafts";
-export * from "./RouteSlipsFollow";
-export * from "./RouteSlipsHistory";
-export * from "./RouteSlipsView";
+export * from "./act";
+export * from "./drafts";
+export * from "./follow";
+export * from "./history";
+export * from "./view";
