+++ front/src/dashboard/slips/view/RouteSlipsView.module.scss
