+++ front/src/dashboard/slips/view/index.ts
@@ -1,0 +1,1 @@
+export * from "./RouteSlipsView";
